Online CV Maker in PHP
I have implemented this project for academic reasons. It is a short project. I have used Water-Fall-Model on (SDLC) to implement this project. In this current time, this web-based application is very useful for all kind of people who need to make a CV in a short time. Anyone can use this system It is very easy to use. Step by step data will be wanted from the user and then It will manage itself and will be given the CV as a "Web page". People can access anyone's CV with a specific URL.
I have used Nodepad++ and xampp server to build this project.